Fix(telestrations text tool): Left and Right arrow keys and back event binding conflicts.
The right and left arrow keys were bound to fast-forward and fast-reverse for the video player.
Thus, if a user wanted to edit text and move the cursor left or right using the keyboard arrows, they
would unexpectedly be moved back or forward in the video. By using Mousetrap, it detects that a textarea
is focused and will disable the events that are bound to the keys for forward/reverse.