Fix(telestrations Cuepoints): Correct the time of the video relative to the telestration time.
Before this change, the video would stop before or after the telestration time and show a noticeable
difference between the scrubber and the cuepoint marker. Instead of emitting pause events from
telestrations, I decided to do time checking in the cuepoints directive. I first check the
cuepoint time within the buffer range, and 'correct' the time by seeking to the cuepoint/telestration time.
After this, the telestrations receive a cuepoint event (new) with the cuepoint time, and show the telestrations
at that time.