fix(video player controls): switch video quality selector to dropdown
This fixes an issue where the select element retained focus after a selection
of one of the options was made causing the video player keyboard controls to not work
until they were clicked on. Using a dropdown instead will not retain focus of the element
and thus the keyboard controls (i.e. spacebar) will still have an immediate effect.