fix(Indexing Playlist): Allows indexer to select events after creating
Adds a new EVENT for when a play is saved to the server. The play
directive intercepts this event and recompiles the play with update play
info from the server. Also, play directive has a local version of the
play (outside of scope) so that Angular doesn't attempt to recompile
template on scope variable change.