fix(xy-coordinate): Adds timeout to initial setting of relative coordinates
Fixes a regression where the marker did not appear on load since the element
it was wrapping did not yet have a height, likely due to the ngModal that
it was wrapped in. The modal had an animation on height/width delaying the
element from having a height in this case. A simple (yet non-ideal) workaround
for now is a slow timeout.