Refactor(upload business logic): Move the business logic of uploads to a service.
Currently, it is necessary to depend on the uploader instance (e.g. Evaporate, Flow)
and the backend to determine the state of an upload since the backend does not push
state or notifications to the frontend. Since we do not know the state of the server
until data is refreshed (page state change/refresh), we need to combine the
uploader instance information with the game video state.

For example, if the user refreshes the page while an upload is in progress,
the server will still be set to 'INCOMPLETE', since it is dumb and doesn't yet
know when an upload has stopped. However, this state essentially means an upload
is in progress. However, since it actually has failed, we need to check if the
uploader instance thinks an upload is in progress at the same time. If there
is no uploader instance for the game OR it doesn't think it is uploading,
then the game is surely not uploading.

This complicated logic should not be present in any template or controller,
since it is common logic to many places interested in the status/state of a game upload.
Therefore, it has been abstracted out into a common service.