refactor(game video status): Move GameVideoStatusService functions into the game.video status functions.
Removed the notion of a GaemVideoStatusService. This is because we want to reduce the complexity of
determining the status of a video upload so a developer only cares about the domain model and doesn't need
to know about some other service. This should reduce future bugs/confusion. Since the functions
isFailed, isIncomplete, & isUploaded, can be wrong because the server is dumb, until it's smarter,
we fill in the gaps/holes of the logic by utilizing the Upload Model if it exists.

Also, we've swapped to using a GUID for tracking a FileUpload, since it's more unique and
more relevant to the file upload itself.