fix(game info): Fix styles so that content has the appropriately set height when page elements dissappear
This is to fix an issue where bottom page content was not taking up all the available space on the page.
The content's height was dependant on other elements being visible on the page, but since some of
these are not conditionally hiding, we need to recalculate these container heights.